dataconnector configs moved to resource
