--- disruptive/dataconnector_configs/__init__.py
@@ -1,1 +1,0 @@
-from disruptive.dataconnector_configs.dataconnector_configs import HttpPush  # noqa
